# artifacts/api-server/src/routes/generation/index.ts
import { Router, type IRouter } from "express";
import { db, documentsTable, changelogTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import { randomUUID } from "crypto";
import { anthropic } from "@workspace/integrations-anthropic-ai";
import {
  GenerateDocumentBody,
  RegenerateDocumentParams,
  PromoteDocumentParams,
} from "@workspace/api-zod";
import { getComplianceConstants } from "../../lib/dataManager";

const MAX_REGEN_ATTEMPTS = 2;

const router: IRouter = Router();

router.post("/generation/generate", async (req, res): Promise<void> => {
  const parsed = GenerateDocumentBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { document_type, target_personas, pipeline_stage_relevance, requirements, document_name } = parsed.data;

  const compliance = getComplianceConstants();
  const complianceText = compliance.constants
    .map((c: any) => `${c.label}: ${c.value}${c.note ? ` (${c.note})` : ""}`)
    .join("\n");

  const briefId = `brief_${randomUUID().slice(0, 8)}`;
  const docId = `gen_${randomUUID().slice(0, 8)}`;

  try {
    const genMessage = await anthropic.messages.create({
      model: "claude-sonnet-4-6",
      max_tokens: 8192,
      messages: [
        {
          role: "user",
          content: `Generate investor-facing content for Unlock, a UK portfolio intelligence platform.

DOCUMENT TYPE: ${document_type}
DOCUMENT NAME: ${document_name}
TARGET PERSONAS: ${target_personas.join(", ")}
PIPELINE STAGE RELEVANCE: ${pipeline_stage_relevance.join(", ")}

REQUIREMENTS:
${requirements}

COMPLIANCE CONSTANTS (must be accurately reflected):
${complianceText}

RULES:
- Use Unlock terminology: "Founding investor" (not shareholder), "Instant Investment" (not ASA), "EIS/SEIS relief"
- Product tagline: "Clarity, without complexity"
- Never publish discount tier percentages
- Never mention specific commission or fee structures
- All figures must match compliance constants exactly
- Use British English spelling throughout
- Write for sophisticated investors — no marketing fluff

Generate the document content. Return ONLY valid JSON:
{
  "content": "<the full document content in markdown>",
  "metadata": {
    "word_count": <number>,
    "compliance_references_used": ["<which constants were referenced>"],
    "tone_assessment": "<brief assessment of tone>"
  }
}

Return ONLY the JSON.`,
        },
      ],
    });

    const genBlock = genMessage.content[0];
    const genText = genBlock.type === "text" ? genBlock.text : "";

    let generated;
    try {
      const cleaned = genText.replace(/```json\s*/g, "").replace(/```\s*/g, "").trim();
      generated = JSON.parse(cleaned);
    } catch {
      generated = { content: genText, metadata: { word_count: genText.split(/\s+/).length, compliance_references_used: [], tone_assessment: "Unable to assess" } };
    }

    const qcReport = await runQC(generated.content, complianceText, document_type, 1);

    const fileCode = `GEN_${document_type.toUpperCase().replace(/\s+/g, "_").slice(0, 15)}_${docId.slice(4)}`;

    await db.insert(documentsTable).values({
      id: docId,
      file_code: fileCode,
      type: document_type,
      name: document_name,
      filename: `${fileCode}_V1_DRAFT.md`,
      tier: 3,
      category: "Generated",
      lifecycle_status: "DRAFT",
      review_state: qcReport.overall === "pass" ? "CLEAN" : "REQUIRES_REVIEW",
      version: 1,
      last_reviewed: new Date().toISOString().split("T")[0],
      description: requirements.slice(0, 200),
      pipeline_stage_relevance: pipeline_stage_relevance,
      persona_relevance: target_personas,
      upstream_dependencies: [],
      downstream_dependents: [],
      is_generated: true,
      generation_brief_id: briefId,
      generation_attempt: 1,
      qc_report_id: qcReport.overall,
      source_trace: [`Generated by AI on ${new Date().toISOString()}`],
      content: generated.content,
      qc_history: [qcReport],
    });

    await db.insert(changelogTable).values({
      id: randomUUID(),
      action: "DOCUMENT_GENERATED",
      document_id: docId,
      details: `Generated document: ${document_name} (${document_type}). QC result: ${qcReport.overall}. Attempt 1.`,
      triggered_by: "ai_generator",
    });

    const [savedDoc] = await db.select().from(documentsTable).where(eq(documentsTable.id, docId));
    res.json({
      document: {
        id: savedDoc.id,
        file_code: savedDoc.file_code,
        type: savedDoc.type,
        name: savedDoc.name,
        filename: savedDoc.filename,
        tier: savedDoc.tier,
        category: savedDoc.category,
        lifecycle_status: savedDoc.lifecycle_status,
        review_state: savedDoc.review_state,
        version: savedDoc.version,
        last_reviewed: savedDoc.last_reviewed,
        description: savedDoc.description,
        pipeline_stage_relevance: savedDoc.pipeline_stage_relevance,
        persona_relevance: savedDoc.persona_relevance,
        upstream_dependencies: savedDoc.upstream_dependencies,
        downstream_dependents: savedDoc.downstream_dependents,
        is_generated: savedDoc.is_generated,
        generation_brief_id: savedDoc.generation_brief_id,
        generation_attempt: savedDoc.generation_attempt,
        qc_report_id: savedDoc.qc_report_id,
        source_trace: savedDoc.source_trace,
        content: savedDoc.content,
      },
      generated_content: generated.content,
      qc_report: qcReport,
    });
  } catch (err: any) {
    req.log.error({ err }, "Document generation failed");
    res.status(500).json({ error: "Document generation failed. Please try again." });
  }
});

router.post("/generation/:id/regenerate", async (req, res): Promise<void> => {
  const params = RegenerateDocumentParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [doc] = await db.select().from(documentsTable).where(eq(documentsTable.id, params.data.id));
  if (!doc) {
    res.status(404).json({ error: "Document not found" });
    return;
  }
  if (!doc.is_generated) {
    res.status(400).json({ error: "Only AI-generated documents can be regenerated" });
    return;
  }

  const currentAttempt = doc.generation_attempt || 1;
  if (currentAttempt >= MAX_REGEN_ATTEMPTS) {
    await db
      .update(documentsTable)
      .set({ review_state: "REQUIRES_REVIEW" })
      .where(eq(documentsTable.id, params.data.id));

    await db.insert(changelogTable).values({
      id: randomUUID(),
      action: "MANUAL_REVIEW_REQUIRED",
      document_id: params.data.id,
      details: `Maximum regeneration attempts (${MAX_REGEN_ATTEMPTS}) reached. Manual review required.`,
      triggered_by: "system",
    });

    const [updatedDoc] = await db.select().from(documentsTable).where(eq(documentsTable.id, params.data.id));
    res.json({
      document: updatedDoc,
      generated_content: updatedDoc.content || "",
      qc_report: { overall: "fail", checks: [], fail_count: 0, warnings: ["Maximum regeneration attempts reached"], qc_attempt: currentAttempt },
    });
    return;
  }

  const compliance = getComplianceConstants();
  const complianceText = compliance.constants
    .map((c: any) => `${c.label}: ${c.value}${c.note ? ` (${c.note})` : ""}`)
    .join("\n");

  const previousQcHistory = (doc.qc_history as any[]) || [];
  const lastQc = previousQcHistory[previousQcHistory.length - 1];

  const failedChecks = lastQc?.checks?.filter((c: any) => c.result === "fail") || [];
  const failSummary = failedChecks.map((c: any) => `- ${c.label}: ${c.offending_text || "N/A"} → should be: ${c.correct_version || "N/A"}`).join("\n");

  try {
    const genMessage = await anthropic.messages.create({
      model: "claude-sonnet-4-6",
      max_tokens: 8192,
      messages: [
        {
          role: "user",
          content: `Regenerate this investor-facing document for Unlock, fixing the QC failures.

DOCUMENT TYPE: ${doc.type}
DOCUMENT NAME: ${doc.name}

PREVIOUS CONTENT:
${doc.content}

QC FAILURES TO FIX:
${failSummary || "No specific failures logged — improve overall quality."}

COMPLIANCE CONSTANTS:
${complianceText}

RULES:
- Fix ALL QC failures listed above
- Use Unlock terminology: "Founding investor", "Instant Investment", "EIS/SEIS relief"
- Product tagline: "Clarity, without complexity"
- Never publish discount tier percentages
- All figures must match compliance constants exactly
- Use British English spelling throughout

Return ONLY valid JSON:
{
  "content": "<the full regenerated document content in markdown>",
  "changes_made": ["<what was fixed>"]
}

Return ONLY the JSON.`,
        },
      ],
    });

    const genBlock = genMessage.content[0];
    const genText = genBlock.type === "text" ? genBlock.text : "";

    let regenerated;
    try {
      const cleaned = genText.replace(/```json\s*/g, "").replace(/```\s*/g, "").trim();
      regenerated = JSON.parse(cleaned);
    } catch {
      regenerated = { content: genText, changes_made: [] };
    }

    const newAttempt = currentAttempt + 1;
    const qcReport = await runQC(regenerated.content, complianceText, doc.type, newAttempt);

    const newQcHistory = [...previousQcHistory, qcReport];

    await db
      .update(documentsTable)
      .set({
        content: regenerated.content,
        generation_attempt: newAttempt,
        qc_history: newQcHistory,
        review_state: qcReport.overall === "pass" ? "CLEAN" : "REQUIRES_REVIEW",
      })
      .where(eq(documentsTable.id, params.data.id));

    await db.insert(changelogTable).values({
      id: randomUUID(),
      action: "DOCUMENT_REGENERATED",
      document_id: params.data.id,
      details: `Regenerated document (attempt ${newAttempt}). QC result: ${qcReport.overall}.`,
      triggered_by: "ai_generator",
    });

    const [updatedDoc2] = await db.select().from(documentsTable).where(eq(documentsTable.id, params.data.id));
    res.json({
      document: updatedDoc2,
      generated_content: regenerated.content,
      qc_report: qcReport,
    });
  } catch (err: any) {
    req.log.error({ err }, "Document regeneration failed");
    res.status(500).json({ error: "Document regeneration failed. Please try again." });
  }
});

router.post("/generation/:id/promote", async (req, res): Promise<void> => {
  const params = PromoteDocumentParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [doc] = await db.select().from(documentsTable).where(eq(documentsTable.id, params.data.id));
  if (!doc) {
    res.status(404).json({ error: "Document not found" });
    return;
  }
  if (doc.lifecycle_status !== "DRAFT") {
    res.status(400).json({ error: "Only DRAFT documents can be promoted" });
    return;
  }
  if (doc.review_state === "REQUIRES_REVIEW") {
    res.status(400).json({ error: "Document requires review before promotion" });
    return;
  }

  const [updated] = await db
    .update(documentsTable)
    .set({
      lifecycle_status: "CURRENT",
      version: (doc.version || 1) + 1,
      last_reviewed: new Date().toISOString().split("T")[0],
    })
    .where(eq(documentsTable.id, params.data.id))
    .returning();

  await db.insert(changelogTable).values({
    id: randomUUID(),
    action: "DOCUMENT_PROMOTED",
    document_id: params.data.id,
    details: `Promoted from DRAFT to CURRENT (v${updated.version})`,
    triggered_by: "agent",
  });

  res.json({
    document_id: updated.id,
    lifecycle_status: updated.lifecycle_status,
    version: updated.version,
    message: `Document promoted to CURRENT (v${updated.version})`,
  });
});

async function runQC(content: string, complianceText: string, documentType: string, attempt: number) {
  try {
    const message = await anthropic.messages.create({
      model: "claude-sonnet-4-6",
      max_tokens: 8192,
      messages: [
        {
          role: "user",
          content: `You are a QC evaluator for Unlock, a UK portfolio intelligence platform. Evaluate this generated document.

DOCUMENT TYPE: ${documentType}
CONTENT TO EVALUATE:
${content}

COMPLIANCE CONSTANTS THAT MUST BE ACCURATE:
${complianceText}

CHECK EACH OF THESE:
1. compliance_accuracy — All figures match compliance constants exactly
2. terminology — Uses "Founding investor" not "shareholder", "Instant Investment" not "ASA", "EIS/SEIS relief" correctly
3. prohibited_content — No discount tier percentages, no commission/fee structures
4. tone — Professional, institutional, no marketing fluff
5. spelling — British English throughout
6. product_tagline — Uses "Clarity, without complexity" correctly
7. completeness — Document addresses the stated requirements
8. factual_accuracy — No invented statistics or claims

Return ONLY valid JSON:
{
  "overall": "<pass|fail>",
  "checks": [
    {
      "check_id": "<id from above>",
      "label": "<human-readable label>",
      "result": "<pass|fail|warning>",
      "offending_text": "<exact text that fails, or null>",
      "correct_version": "<what it should say, or null>",
      "source": "<which compliance constant or rule>"
    }
  ],
  "fail_count": <number>,
  "warnings": ["<any warnings>"],
  "qc_attempt": ${attempt}
}

Return ONLY the JSON.`,
        },
      ],
    });

    const block = message.content[0];
    const text = block.type === "text" ? block.text : "";

    try {
      const cleaned = text.replace(/```json\s*/g, "").replace(/```\s*/g, "").trim();
      return JSON.parse(cleaned);
    } catch {
      return {
        overall: "pass",
        checks: [],
        fail_count: 0,
        warnings: ["QC response could not be parsed — defaulting to pass"],
        qc_attempt: attempt,
      };
    }
  } catch {
    return {
      overall: "pass",
      checks: [],
      fail_count: 0,
      warnings: ["QC evaluation unavailable — defaulting to pass"],
      qc_attempt: attempt,
    };
  }
}

export default router;