# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Unlock Content Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /generation/generate:
    post:
      operationId: generateDocument
      tags: [generation]
      summary: Generate a document from a brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerationBrief"
      responses:
        "200":
          description: Generated document with QC report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResult"
  /generation/{id}/regenerate:
    post:
      operationId: regenerateDocument
      tags: [generation]
      summary: Regenerate a document with QC failures as constraints
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Regeneration result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResult"
  /generation/{id}/promote:
    post:
      operationId: promoteDocument
      tags: [generation]
      summary: Promote a DRAFT document to CURRENT
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Promoted document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
...
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
        file_code:
          type: string
        type:
          type: string
        name:
          type: string
        filename:
          type: string
        tier:
          type: integer
        category:
          type: string
        lifecycle_status:
          type: string
        review_state:
          type: string
        version:
          type: integer
        last_reviewed:
          type: string
        description:
          type: string
        pipeline_stage_relevance:
          type: array
          items:
            type: string
        persona_relevance:
          type: array
          items:
            type: string
        upstream_dependencies:
          type: array
          items:
            type: string
        downstream_dependents:
          type: array
          items:
            type: string
        is_generated:
          type: boolean
        generation_brief_id:
          type: ["string", "null"]
        generation_attempt:
          type: ["integer", "null"]
        qc_report_id:
          type: ["string", "null"]
        source_trace:
          type: array
          items:
            $ref: "#/components/schemas/SourceTraceEntry"
        content:
          type: ["string", "null"]
      required:
        - id
        - file_code
        - type
        - name
        - filename
        - tier
        - category
        - lifecycle_status
        - review_state
        - version
        - last_reviewed
        - description
        - pipeline_stage_relevance
        - persona_relevance
        - upstream_dependencies
        - downstream_dependents
        - is_generated
        - source_trace
    SourceTraceEntry:
      type: object
      properties:
        claim_area:
          type: string
        source_document_id:
          type: string
        source_section:
          type: string
        source_excerpt_key:
          type: string
      required:
        - claim_area
        - source_document_id
        - source_section
        - source_excerpt_key
...
    GenerationBrief:
      type: object
      properties:
        document_type:
          type: string
        target_personas:
          type: array
          items:
            type: string
        pipeline_stage_relevance:
          type: array
          items:
            type: string
        requirements:
          type: string
        document_name:
          type: string
      required:
        - document_type
        - target_personas
        - pipeline_stage_relevance
        - requirements
        - document_name
    GenerationResult:
      type: object
      properties:
        document:
          $ref: "#/components/schemas/Document"
        generated_content:
          type: string
        qc_report:
          $ref: "#/components/schemas/QCReport"
      required:
        - document
        - generated_content
        - qc_report
    QCReport:
      type: object
      properties:
        overall:
          type: string
        checks:
          type: array
          items:
            $ref: "#/components/schemas/QCCheck"
        fail_count:
          type: integer
        warnings:
          type: array
          items:
            type: string
        qc_attempt:
          type: integer
      required:
        - overall
        - checks
        - fail_count
        - warnings
        - qc_attempt
    QCCheck:
      type: object
      properties:
        check_id:
          type: string
        label:
          type: string
        result:
          type: string
        offending_text:
          type: ["string", "null"]
        correct_version:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
      required:
        - check_id
        - label
        - result
...